Extract individual measures into their own classes, implement school comparison headline measures
