+++ SAPSec.Core/Features/Ks4HeadlineMeasures/UseCases/GetSchoolComparisonKs4HeadlineMeasures.cs
@@ -1,4 +1,3 @@
-using SAPSec.Core.Features.Ks4HeadlineMeasures;
 using SAPSec.Core.Features.Measures;
 using SAPSec.Core.Features.SimilarSchools;
 using SAPSec.Core.Interfaces.Repositories;
@@ -6,15 +5,16 @@ using SAPSec.Core.Interfaces.Services;
 using SAPSec.Core.Model;
 using SAPSec.Core.Model.Generated;
 
-namespace SAPSec.Core.Features.Ks4CoreSubjects.UseCases;
+namespace SAPSec.Core.Features.Ks4HeadlineMeasures.UseCases;
 
-public class GetSchoolComparisonKs4CoreSubjects(
+public class GetSchoolComparisonKs4HeadlineMeasures(
     IKs4PerformanceRepository performanceRepository,
+    IKs4DestinationsRepository destinationsRepository,
     ISchoolDetailsService schoolDetailsService,
     IEstablishmentRepository establishmentRepository,
     ISimilarSchoolsSecondaryRepository similarSchoolsRepository)
 {
-    public async Task<GetSchoolComparisonKs4CoreSubjectsResponse> Execute(GetSchoolComparisonKs4CoreSubjectsRequest request)
+    public async Task<GetSchoolComparisonKs4HeadlineMeasuresResponse> Execute(GetSchoolComparisonKs4HeadlineMeasuresRequest request)
     {
         var currentSchoolDetails = await schoolDetailsService.GetByUrnAsync(request.CurrentSchoolUrn);
         var similarSchoolDetails = await schoolDetailsService.GetByUrnAsync(request.SimilarSchoolUrn);
@@ -23,7 +23,7 @@ public class GetSchoolComparisonKs4CoreSubjects(
             request.CurrentSchoolUrn,
             currentSchoolDetails.Name,
             await performanceRepository.GetByUrnAsync(request.CurrentSchoolUrn),
-            null);
+            await destinationsRepository.GetByUrnAsync(request.CurrentSchoolUrn));
 
         var similarSchoolsUrns = (await similarSchoolsRepository.GetSimilarSchoolsGroupAsync(request.CurrentSchoolUrn))
             .Select(g => g.NeighbourURN)
@@ -31,10 +31,13 @@ public class GetSchoolComparisonKs4CoreSubjects(
             .Distinct(StringComparer.Ordinal)
             .ToArray();
 
-        var similarSchoolsPerformance = ((await performanceRepository.GetByUrnsAsync(similarSchoolsUrns)) ?? [])
+        var similarSchoolsPerformance = (await performanceRepository.GetByUrnsAsync(similarSchoolsUrns) ?? [])
             .ToDictionary(x => x.URN, x => x, StringComparer.Ordinal);
 
-        var similarSchoolsDetails = ((await establishmentRepository.GetEstablishmentsAsync(similarSchoolsUrns))
+        var similarSchoolsDestinations = (await destinationsRepository.GetByUrnsAsync(similarSchoolsUrns) ?? [])
+            .ToDictionary(x => x.Urn, x => x, StringComparer.Ordinal);
+
+        var similarSchoolsDetails = (await establishmentRepository.GetEstablishmentsAsync(similarSchoolsUrns)
                 ?? Array.Empty<Establishment>())
             .Where(x => !string.IsNullOrWhiteSpace(x.URN))
             .ToDictionary(x => x.URN, StringComparer.Ordinal);
@@ -45,7 +48,7 @@ public class GetSchoolComparisonKs4CoreSubjects(
                 urn,
                 similarSchoolsDetails[urn].EstablishmentName,
                 similarSchoolsPerformance.GetValueOrDefault(urn),
-                null))
+                similarSchoolsDestinations.GetValueOrDefault(urn)))
             .ToDictionary(x => x.Urn, x => x, StringComparer.Ordinal);
 
         var similarSchoolData = similarSchoolsData[request.SimilarSchoolUrn];
@@ -56,25 +59,21 @@ public class GetSchoolComparisonKs4CoreSubjects(
             currentSchoolDetails,
             similarSchoolDetails,
             similarSchoolsData.Values.Count,
-            [
-                Measures.Ks4CoreSubjects.EnglishLanguage.ForSchoolComparison(currentSchoolData, similarSchoolData, similarSchoolsData, filterBy),
-                Measures.Ks4CoreSubjects.EnglishLiterature.ForSchoolComparison(currentSchoolData, similarSchoolData, similarSchoolsData, filterBy),
-                Measures.Ks4CoreSubjects.Biology.ForSchoolComparison(currentSchoolData, similarSchoolData, similarSchoolsData, filterBy),
-                Measures.Ks4CoreSubjects.Chemistry.ForSchoolComparison(currentSchoolData, similarSchoolData, similarSchoolsData, filterBy),
-                Measures.Ks4CoreSubjects.Physics.ForSchoolComparison(currentSchoolData, similarSchoolData, similarSchoolsData, filterBy),
-                Measures.Ks4CoreSubjects.Mathematics.ForSchoolComparison(currentSchoolData, similarSchoolData, similarSchoolsData, filterBy),
-                Measures.Ks4CoreSubjects.CombinedScience.ForSchoolComparison(currentSchoolData, similarSchoolData, similarSchoolsData, filterBy),
-            ]);
+            Measures.Ks4HeadlineMeasures.Attainment8.ForSchoolComparison(currentSchoolData, similarSchoolData, similarSchoolsData, filterBy),
+            Measures.Ks4HeadlineMeasures.EnglishAndMaths.ForSchoolComparison(currentSchoolData, similarSchoolData, similarSchoolsData, filterBy),
+            Measures.Ks4HeadlineMeasures.Destinations.ForSchoolComparison(currentSchoolData, similarSchoolData, similarSchoolsData, filterBy));
     }
 }
 
-public record GetSchoolComparisonKs4CoreSubjectsRequest(
+public record GetSchoolComparisonKs4HeadlineMeasuresRequest(
     string CurrentSchoolUrn,
     string SimilarSchoolUrn,
     IDictionary<string, string>? FilterBy = null);
 
-public record GetSchoolComparisonKs4CoreSubjectsResponse(
+public record GetSchoolComparisonKs4HeadlineMeasuresResponse(
     SchoolDetails CurrentSchoolDetails,
     SchoolDetails SimilarSchoolDetails,
     int SimilarSchoolsCount,
-    IReadOnlyCollection<Measure> Measures);
+    Measure Attainment8,
+    Measure EnglishAndMaths,
+    Measure Destinations);
