+++ SAPSec.Core/Features/Measures/SubMeasure.cs
@@ -1,0 +1,5 @@
+namespace SAPSec.Core.Features.Measures;
+
+//////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+public abstract record SubMeasure;
