+++ SAPSec.Web/Controllers/SimilarSchoolsComparisonController.cs
@@ -3,11 +3,13 @@ using Microsoft.AspNetCore.Mvc;
 using SAPSec.Core.Features.Attendance.UseCases;
 using SAPSec.Core.Features.Ks4CoreSubjects.UseCases;
 using SAPSec.Core.Features.Ks4HeadlineMeasures.UseCases;
+using SAPSec.Core.Features.Measures;
 using SAPSec.Core.Features.SimilarSchools.UseCases;
 using SAPSec.Core.Model;
 using SAPSec.Web.Constants;
 using SAPSec.Web.Formatters;
 using SAPSec.Web.ViewModels;
+using SAPSec.Web.ViewModels.Measures;
 using System.Globalization;
 
 namespace SAPSec.Web.Controllers;
@@ -19,7 +21,7 @@ public class SimilarSchoolsComparisonController : Controller
     private readonly GetSimilarSchoolDetails _getSimilarSchoolDetails;
     private readonly GetAttendanceMeasures _getAttendanceMeasures;
     private readonly GetSchoolComparisonKs4CoreSubjects _getSchoolComparisonKs4CoreSubjects;
-    private readonly GetKs4HeadlineMeasures _getKs4HeadlineMeasures;
+    private readonly GetSchoolComparisonKs4HeadlineMeasures _getSchoolComparisonKs4HeadlineMeasures;
     private readonly GetCharacteristicsComparison _getCharacteristicsComparison;
     private readonly ILogger<SimilarSchoolsComparisonController> _logger;
     private readonly ICharacteristicsComparisonFormatter _characteristicsFormatter;
@@ -28,21 +30,25 @@ public class SimilarSchoolsComparisonController : Controller
         GetSimilarSchoolDetails getSimilarSchoolDetails,
         GetAttendanceMeasures getAttendanceMeasures,
         GetSchoolComparisonKs4CoreSubjects getSchoolComparisonKs4CoreSubjects,
-        GetKs4HeadlineMeasures getKs4HeadlineMeasures,
+        GetSchoolComparisonKs4HeadlineMeasures getSchoolComparisonKs4HeadlineMeasures,
         GetCharacteristicsComparison getCharacteristicsComparison,
         ICharacteristicsComparisonFormatter characteristicsFormatter,
         ILogger<SimilarSchoolsComparisonController> logger)
     {
-        _getSimilarSchoolDetails =
-            getSimilarSchoolDetails ?? throw new ArgumentNullException(nameof(getSimilarSchoolDetails));
-        _getAttendanceMeasures = getAttendanceMeasures ?? throw new ArgumentNullException(nameof(getAttendanceMeasures));
-        _getSchoolComparisonKs4CoreSubjects = getSchoolComparisonKs4CoreSubjects ?? throw new ArgumentNullException(nameof(getSchoolComparisonKs4CoreSubjects));
-        _getKs4HeadlineMeasures = getKs4HeadlineMeasures ?? throw new ArgumentNullException(nameof(getKs4HeadlineMeasures));
-        _getCharacteristicsComparison = getCharacteristicsComparison ??
-                                        throw new ArgumentNullException(nameof(getCharacteristicsComparison));
-        _characteristicsFormatter = characteristicsFormatter ??
-                                    throw new ArgumentNullException(nameof(characteristicsFormatter));
-        _logger = logger ?? throw new ArgumentNullException(nameof(logger));
+        _getSimilarSchoolDetails = getSimilarSchoolDetails
+            ?? throw new ArgumentNullException(nameof(getSimilarSchoolDetails));
+        _getAttendanceMeasures = getAttendanceMeasures
+            ?? throw new ArgumentNullException(nameof(getAttendanceMeasures));
+        _getSchoolComparisonKs4CoreSubjects = getSchoolComparisonKs4CoreSubjects
+            ?? throw new ArgumentNullException(nameof(getSchoolComparisonKs4CoreSubjects));
+        _getSchoolComparisonKs4HeadlineMeasures = getSchoolComparisonKs4HeadlineMeasures
+            ?? throw new ArgumentNullException(nameof(getSchoolComparisonKs4HeadlineMeasures));
+        _getCharacteristicsComparison = getCharacteristicsComparison
+            ?? throw new ArgumentNullException(nameof(getCharacteristicsComparison));
+        _characteristicsFormatter = characteristicsFormatter
+            ?? throw new ArgumentNullException(nameof(characteristicsFormatter));
+        _logger = logger
+            ?? throw new ArgumentNullException(nameof(logger));
     }
 
     [HttpGet]
@@ -53,7 +59,7 @@ public class SimilarSchoolsComparisonController : Controller
         Similarity(urn, similarSchoolUrn, similarityCalculation);
 
     [HttpGet]
-    [Route("Similarity")]
+    [Route("similarity")]
     public async Task<IActionResult> Similarity(
         string urn,
         string similarSchoolUrn,
@@ -70,7 +76,7 @@ public class SimilarSchoolsComparisonController : Controller
     }
 
     [HttpGet]
-    [Route("Ks4HeadlineMeasures")]
+    [Route("ks4-headline-measures")]
     public async Task<IActionResult> Ks4HeadlineMeasures(
         string urn,
         string similarSchoolUrn,
@@ -80,227 +86,17 @@ public class SimilarSchoolsComparisonController : Controller
         if (modelResult.Result != null)
             return modelResult.Result;
 
-        var thisSchoolKs4 = await _getKs4HeadlineMeasures.Execute(new GetKs4HeadlineMeasuresRequest(urn));
-        var selectedSchoolKs4 = await _getKs4HeadlineMeasures.Execute(new GetKs4HeadlineMeasuresRequest(similarSchoolUrn));
+        var filters = Request.Query.ToDictionary(r => r.Key, r => r.Value.ToString());
+        var response = await _getSchoolComparisonKs4HeadlineMeasures.Execute(new GetSchoolComparisonKs4HeadlineMeasuresRequest(urn, similarSchoolUrn, filters));
+        ViewData[ViewDataKeys.BreadcrumbNode] = BreadcrumbNodes.SchoolHome(urn);
 
         var model = modelResult.Model!;
-        model.ThisSchoolAttainment8ThreeYearAverage = thisSchoolKs4?.Attainment8ThreeYearAverage.SchoolValue;
-        model.SelectedSchoolAttainment8ThreeYearAverage = selectedSchoolKs4?.Attainment8ThreeYearAverage.SchoolValue;
-        model.EnglandAttainment8ThreeYearAverage =
-            thisSchoolKs4?.Attainment8ThreeYearAverage.EnglandValue
-            ?? selectedSchoolKs4?.Attainment8ThreeYearAverage.EnglandValue;
-        model.ThisSchoolAttainment8YearByYear = thisSchoolKs4?.Attainment8YearByYear.School;
-        model.SelectedSchoolAttainment8YearByYear = selectedSchoolKs4?.Attainment8YearByYear.School;
-        model.EnglandAttainment8YearByYear =
-            thisSchoolKs4?.Attainment8YearByYear.England
-            ?? selectedSchoolKs4?.Attainment8YearByYear.England;
-        model.ThisSchoolEngMaths49ThreeYearAverage = thisSchoolKs4?.EngMaths49ThreeYearAverage.SchoolValue;
-        model.SelectedSchoolEngMaths49ThreeYearAverage = selectedSchoolKs4?.EngMaths49ThreeYearAverage.SchoolValue;
-        model.EnglandEngMaths49ThreeYearAverage =
-            thisSchoolKs4?.EngMaths49ThreeYearAverage.EnglandValue
-            ?? selectedSchoolKs4?.EngMaths49ThreeYearAverage.EnglandValue;
-        model.ThisSchoolEngMaths49YearByYear = thisSchoolKs4?.EngMaths49YearByYear.School;
-        model.SelectedSchoolEngMaths49YearByYear = selectedSchoolKs4?.EngMaths49YearByYear.School;
-        model.EnglandEngMaths49YearByYear =
-            thisSchoolKs4?.EngMaths49YearByYear.England
-            ?? selectedSchoolKs4?.EngMaths49YearByYear.England;
-        model.ThisSchoolEngMaths59ThreeYearAverage = thisSchoolKs4?.EngMaths59ThreeYearAverage.SchoolValue;
-        model.SelectedSchoolEngMaths59ThreeYearAverage = selectedSchoolKs4?.EngMaths59ThreeYearAverage.SchoolValue;
-        model.EnglandEngMaths59ThreeYearAverage =
-            thisSchoolKs4?.EngMaths59ThreeYearAverage.EnglandValue
-            ?? selectedSchoolKs4?.EngMaths59ThreeYearAverage.EnglandValue;
-        model.ThisSchoolEngMaths59YearByYear = thisSchoolKs4?.EngMaths59YearByYear.School;
-        model.SelectedSchoolEngMaths59YearByYear = selectedSchoolKs4?.EngMaths59YearByYear.School;
-        model.EnglandEngMaths59YearByYear =
-            thisSchoolKs4?.EngMaths59YearByYear.England
-            ?? selectedSchoolKs4?.EngMaths59YearByYear.England;
-        model.ThisSchoolDestinationsThreeYearAverage = thisSchoolKs4?.DestinationsThreeYearAverage.SchoolValue;
-        model.SelectedSchoolDestinationsThreeYearAverage = selectedSchoolKs4?.DestinationsThreeYearAverage.SchoolValue;
-        model.EnglandDestinationsThreeYearAverage =
-            thisSchoolKs4?.DestinationsThreeYearAverage.EnglandValue
-            ?? selectedSchoolKs4?.DestinationsThreeYearAverage.EnglandValue;
-        model.ThisSchoolDestinationsYearByYear = thisSchoolKs4?.DestinationsYearByYear.School;
-        model.SelectedSchoolDestinationsYearByYear = selectedSchoolKs4?.DestinationsYearByYear.School;
-        model.EnglandDestinationsYearByYear =
-            thisSchoolKs4?.DestinationsYearByYear.England
-            ?? selectedSchoolKs4?.DestinationsYearByYear.England;
-
+        model.Attainment8 = MapMeasure(response.Attainment8, response.CurrentSchoolDetails, response.SimilarSchoolDetails);
+        model.EnglishAndMaths = MapMeasure(response.EnglishAndMaths, response.CurrentSchoolDetails, response.SimilarSchoolDetails);
+        model.Destinations = MapMeasure(response.Destinations, response.CurrentSchoolDetails, response.SimilarSchoolDetails);
         SetComparisonSchoolViewData(model);
+
         return View(model);
-    }
-
-    [HttpGet]
-    [Route("Ks4HeadlineMeasuresData")]
-    public async Task<IActionResult> Ks4HeadlineMeasuresData(string urn, string similarSchoolUrn, string grade = "4")
-    {
-        if (string.IsNullOrWhiteSpace(urn) || string.IsNullOrWhiteSpace(similarSchoolUrn))
-        {
-            return BadRequest(new { error = "Missing route parameters." });
-        }
-
-        var normalizedGrade = grade == "5" ? "5" : "4";
-        var thisSchoolKs4 = await _getKs4HeadlineMeasures.Execute(new GetKs4HeadlineMeasuresRequest(urn));
-        var selectedSchoolKs4 = await _getKs4HeadlineMeasures.Execute(new GetKs4HeadlineMeasuresRequest(similarSchoolUrn));
-
-        var isGrade5 = normalizedGrade == "5";
-
-        var thisSchoolThreeYear = isGrade5
-            ? thisSchoolKs4?.EngMaths59ThreeYearAverage.SchoolValue
-            : thisSchoolKs4?.EngMaths49ThreeYearAverage.SchoolValue;
-        var selectedSchoolThreeYear = isGrade5
-            ? selectedSchoolKs4?.EngMaths59ThreeYearAverage.SchoolValue
-            : selectedSchoolKs4?.EngMaths49ThreeYearAverage.SchoolValue;
-        var englandThreeYear = isGrade5
-            ? (thisSchoolKs4?.EngMaths59ThreeYearAverage.EnglandValue ?? selectedSchoolKs4?.EngMaths59ThreeYearAverage.EnglandValue)
-            : (thisSchoolKs4?.EngMaths49ThreeYearAverage.EnglandValue ?? selectedSchoolKs4?.EngMaths49ThreeYearAverage.EnglandValue);
-
-        var thisSchoolSeries = isGrade5
-            ? thisSchoolKs4?.EngMaths59YearByYear.School
-            : thisSchoolKs4?.EngMaths49YearByYear.School;
-        var selectedSchoolSeries = isGrade5
-            ? selectedSchoolKs4?.EngMaths59YearByYear.School
-            : selectedSchoolKs4?.EngMaths49YearByYear.School;
-        var englandSeries = isGrade5
-            ? (thisSchoolKs4?.EngMaths59YearByYear.England ?? selectedSchoolKs4?.EngMaths59YearByYear.England)
-            : (thisSchoolKs4?.EngMaths49YearByYear.England ?? selectedSchoolKs4?.EngMaths49YearByYear.England);
-
-        return Json(new
-        {
-            grade = normalizedGrade,
-            bar = new decimal?[]
-            {
-                thisSchoolThreeYear,
-                selectedSchoolThreeYear,
-                englandThreeYear
-            },
-            line = new
-            {
-                thisSchool = new decimal?[]
-                {
-                    thisSchoolSeries?.Previous2,
-                    thisSchoolSeries?.Previous,
-                    thisSchoolSeries?.Current
-                },
-                similarSchool = new decimal?[]
-                {
-                    selectedSchoolSeries?.Previous2,
-                    selectedSchoolSeries?.Previous,
-                    selectedSchoolSeries?.Current
-                },
-                england = new decimal?[]
-                {
-                    englandSeries?.Previous2,
-                    englandSeries?.Previous,
-                    englandSeries?.Current
-                }
-            },
-            table = new
-            {
-                thisSchool = new[]
-                {
-                    SimilarSchoolsComparisonViewModel.DisplayWholePercent(thisSchoolSeries?.Previous2),
-                    SimilarSchoolsComparisonViewModel.DisplayWholePercent(thisSchoolSeries?.Previous),
-                    SimilarSchoolsComparisonViewModel.DisplayWholePercent(thisSchoolSeries?.Current),
-                    SimilarSchoolsComparisonViewModel.DisplayWholePercent(thisSchoolThreeYear)
-                },
-                similarSchool = new[]
-                {
-                    SimilarSchoolsComparisonViewModel.DisplayWholePercent(selectedSchoolSeries?.Previous2),
-                    SimilarSchoolsComparisonViewModel.DisplayWholePercent(selectedSchoolSeries?.Previous),
-                    SimilarSchoolsComparisonViewModel.DisplayWholePercent(selectedSchoolSeries?.Current),
-                    SimilarSchoolsComparisonViewModel.DisplayWholePercent(selectedSchoolThreeYear)
-                },
-                england = new[]
-                {
-                    SimilarSchoolsComparisonViewModel.DisplayWholePercent(englandSeries?.Previous2),
-                    SimilarSchoolsComparisonViewModel.DisplayWholePercent(englandSeries?.Previous),
-                    SimilarSchoolsComparisonViewModel.DisplayWholePercent(englandSeries?.Current),
-                    SimilarSchoolsComparisonViewModel.DisplayWholePercent(englandThreeYear)
-                }
-            }
-        });
-    }
-
-    [HttpGet]
-    [Route("Ks4DestinationsData")]
-    public async Task<IActionResult> Ks4DestinationsData(string urn, string similarSchoolUrn, string destination = "all")
-    {
-        if (string.IsNullOrWhiteSpace(urn) || string.IsNullOrWhiteSpace(similarSchoolUrn))
-        {
-            return BadRequest(new { error = "Missing route parameters." });
-        }
-
-        var normalizedDestination = NormalizeDestinationFilter(destination);
-        var thisSchoolKs4 = await _getKs4HeadlineMeasures.Execute(new GetKs4HeadlineMeasuresRequest(urn));
-        var selectedSchoolKs4 = await _getKs4HeadlineMeasures.Execute(new GetKs4HeadlineMeasuresRequest(similarSchoolUrn));
-
-        var thisSchoolThreeYear = SelectDestinationsAverage(thisSchoolKs4, normalizedDestination)?.SchoolValue;
-        var selectedSchoolThreeYear = SelectDestinationsAverage(selectedSchoolKs4, normalizedDestination)?.SchoolValue;
-        var englandThreeYear =
-            SelectDestinationsAverage(thisSchoolKs4, normalizedDestination)?.EnglandValue
-            ?? SelectDestinationsAverage(selectedSchoolKs4, normalizedDestination)?.EnglandValue;
-
-        var thisSchoolSeries = SelectDestinationsYearByYear(thisSchoolKs4, normalizedDestination)?.School;
-        var selectedSchoolSeries = SelectDestinationsYearByYear(selectedSchoolKs4, normalizedDestination)?.School;
-        var englandSeries =
-            SelectDestinationsYearByYear(thisSchoolKs4, normalizedDestination)?.England
-            ?? SelectDestinationsYearByYear(selectedSchoolKs4, normalizedDestination)?.England;
-
-        return Json(new
-        {
-            destination = normalizedDestination,
-            bar = new decimal?[]
-            {
-                thisSchoolThreeYear,
-                selectedSchoolThreeYear,
-                englandThreeYear
-            },
-            line = new
-            {
-                thisSchool = new decimal?[]
-                {
-                    thisSchoolSeries?.Previous2,
-                    thisSchoolSeries?.Previous,
-                    thisSchoolSeries?.Current
-                },
-                similarSchool = new decimal?[]
-                {
-                    selectedSchoolSeries?.Previous2,
-                    selectedSchoolSeries?.Previous,
-                    selectedSchoolSeries?.Current
-                },
-                england = new decimal?[]
-                {
-                    englandSeries?.Previous2,
-                    englandSeries?.Previous,
-                    englandSeries?.Current
-                }
-            },
-            table = new
-            {
-                thisSchool = new[]
-                {
-                    SimilarSchoolsComparisonViewModel.DisplayWholePercent(thisSchoolSeries?.Previous2),
-                    SimilarSchoolsComparisonViewModel.DisplayWholePercent(thisSchoolSeries?.Previous),
-                    SimilarSchoolsComparisonViewModel.DisplayWholePercent(thisSchoolSeries?.Current),
-                    SimilarSchoolsComparisonViewModel.DisplayWholePercent(thisSchoolThreeYear)
-                },
-                similarSchool = new[]
-                {
-                    SimilarSchoolsComparisonViewModel.DisplayWholePercent(selectedSchoolSeries?.Previous2),
-                    SimilarSchoolsComparisonViewModel.DisplayWholePercent(selectedSchoolSeries?.Previous),
-                    SimilarSchoolsComparisonViewModel.DisplayWholePercent(selectedSchoolSeries?.Current),
-                    SimilarSchoolsComparisonViewModel.DisplayWholePercent(selectedSchoolThreeYear)
-                },
-                england = new[]
-                {
-                    SimilarSchoolsComparisonViewModel.DisplayWholePercent(englandSeries?.Previous2),
-                    SimilarSchoolsComparisonViewModel.DisplayWholePercent(englandSeries?.Previous),
-                    SimilarSchoolsComparisonViewModel.DisplayWholePercent(englandSeries?.Current),
-                    SimilarSchoolsComparisonViewModel.DisplayWholePercent(englandThreeYear)
-                }
-            }
-        });
     }
 
     [HttpGet]
@@ -324,61 +120,6 @@ public class SimilarSchoolsComparisonController : Controller
 
         return View(model);
     }
-
-    private static MeasureViewModel MapMeasure(SchoolComparisonMeasure measure, SchoolDetails currentSchoolDetails, SchoolDetails similarSchoolDetails)
-    {
-        var chartColors = new[] { "#ca357c", "#2a1950", "#2a1950" };
-        var yearByYearColors = new[] { "#ca357c", "#2a1950", "#4b9b7d" };
-
-        var measureInfo = new MeasureInfoViewModel(
-            measure.Key,
-            measure.Name,
-            measure.DataType,
-            measure.AvailableFilters.Select(MapAvailableFilter),
-            [
-                currentSchoolDetails.Name,
-                similarSchoolDetails.Name,
-                "Schools in England average",
-            ]);
-
-        return new(measureInfo, [
-            new ThreeYearAverageSubMeasureViewModel(
-                "three-year-average",
-                "3-year average",
-                measureInfo,
-                [
-                    measure.ThreeYearAverage.CurrentSchoolValue,
-                    measure.ThreeYearAverage.SimilarSchoolValue,
-                    measure.ThreeYearAverage.EnglandValue,
-                ],
-                chartColors),
-            new YearByYearSubMeasureViewModel(
-                "year-by-year",
-                "Year by year",
-                measureInfo,
-                [
-                    MapYearByYear(measure.YearByYear.CurrentSchool),
-                    MapYearByYear(measure.YearByYear.SimilarSchool),
-                    MapYearByYear(measure.YearByYear.England),
-                ],
-                yearByYearColors),
-            new TableSubMeasureViewModel(
-                "table-view",
-                "Table",
-                measureInfo,
-                [
-                    new("this", MapYearByYear(measure.YearByYear.CurrentSchool), measure.ThreeYearAverage.CurrentSchoolValue),
-                    new("similar", MapYearByYear(measure.YearByYear.SimilarSchool), measure.ThreeYearAverage.SimilarSchoolValue),
-                    new("england", MapYearByYear(measure.YearByYear.England), measure.ThreeYearAverage.EnglandValue),
-                ])
-        ]);
-    }
-
-    private static MeasureAvailableFilterViewModel MapAvailableFilter(MeasureAvailableFilter availableFilter) =>
-        new(availableFilter.Key, availableFilter.Name, availableFilter.Options.Select(o => new MeasureFilterOptionViewModel(o.Key, o.Name, o.Count, o.Selected)));
-
-    private static YearByYearSeriesViewModel MapYearByYear(MeasureYearByYearSeries yearByYear) =>
-        new(yearByYear.Current, yearByYear.Previous, yearByYear.Previous2);
 
     [HttpGet]
     [Route("attendance")]
@@ -479,7 +220,7 @@ public class SimilarSchoolsComparisonController : Controller
     }
 
     [HttpGet]
-    [Route("SchoolDetails")]
+    [Route("school-details")]
     public async Task<IActionResult> SchoolDetails(
         string urn,
         string similarSchoolUrn,
@@ -491,6 +232,23 @@ public class SimilarSchoolsComparisonController : Controller
 
         SetComparisonSchoolViewData(modelResult.Model!);
         return View(modelResult.Model);
+    }
+
+    private static MeasureViewModel MapMeasure(Measure measure, SchoolDetails currentSchoolDetails, SchoolDetails similarSchoolDetails)
+    {
+        // TODO: Should colours be in the view? Not sure how to do that yet.
+        var chartColors = new[] { "#ca357c", "#2a1950", "#2a1950" };
+        var yearByYearColors = new[] { "#ca357c", "#2a1950", "#4b9b7d" };
+
+        return MeasureViewModel.FromMeasure(measure, currentSchoolDetails,
+            // TODO: Should labels be in the view? Not sure how to do that yet.
+            [
+                currentSchoolDetails.Name,
+                similarSchoolDetails.Name,
+                "Schools in England average",
+            ],
+            chartColors,
+            yearByYearColors);
     }
 
     /// <summary>
@@ -611,29 +369,4 @@ public class SimilarSchoolsComparisonController : Controller
         value.HasValue
             ? value.Value.ToString("0.00", CultureInfo.InvariantCulture) + "%"
             : "No available data";
-
-    private static string NormalizeDestinationFilter(string? destination) =>
-        destination?.ToLowerInvariant() switch
-        {
-            "education" => "education",
-            "employment" => "employment",
-            _ => "all"
-        };
-
-    private static Ks4HeadlineMeasureAverage? SelectDestinationsAverage(GetKs4HeadlineMeasuresResponse? response, string destination) =>
-        destination switch
-        {
-            "education" => response?.DestinationsEducationThreeYearAverage,
-            "employment" => response?.DestinationsEmploymentThreeYearAverage,
-            _ => response?.DestinationsThreeYearAverage
-        };
-
-    private static Ks4HeadlineMeasureYearByYear? SelectDestinationsYearByYear(GetKs4HeadlineMeasuresResponse? response, string destination) =>
-        destination switch
-        {
-            "education" => response?.DestinationsEducationYearByYear,
-            "employment" => response?.DestinationsEmploymentYearByYear,
-            _ => response?.DestinationsYearByYear
-        };
-
 }
