+++ SAPSec.Web/Views/School/Ks4HeadlineMeasures.cshtml
@@ -44,19 +44,7 @@
     <section aria-labelledby="destinations-heading" class="app-ks4-attainment govuk-!-margin-top-9">
         <h2 class="govuk-heading-m" id="destinations-heading">Staying in education or entering employment</h2>
 
-        <details class="govuk-details app-ks4-copy govuk-!-margin-bottom-3 app-ks4-details" data-module="govuk-details">
-            <summary class="govuk-details__summary">
-                <span class="govuk-details__summary-text">Information about staying in education or entering employment data</span>
-            </summary>
-            <div class="govuk-details__text">
-                <p class="govuk-body-s govuk-!-margin-bottom-2">
-                    This data shows the number of pupils who either stayed in education, started an apprenticeship, or entered employment for at least two terms after completing key stage 4.
-                </p>
-                <p class="govuk-body-s govuk-!-margin-bottom-2">
-                    The most recent data covers pupils who finished year 11 in 2023. The 3-year average is the average of pupils completed year 11 in 2021,2022 and 2023.
-                </p>
-            </div>
-        </details>
+        <partial name="~/Views/Shared/_Ks4DestinationsIntro.cshtml" />
 
         <partial name="_Measure" model=@Model.Destinations />
     </section>
