+++ SAPSec.Web/Views/Shared/_Measure.cshtml
@@ -1,6 +1,6 @@
 ﻿@using SAPSec.Web.Constants
 @using System.Text.Json
-@using SAPSec.Web.ViewModels
+@using SAPSec.Web.ViewModels.Measures
 
 @model MeasureViewModel
