+++ SAPSec.Web/Views/Shared/_MeasureFilters.cshtml
@@ -1,4 +1,4 @@
-﻿@using SAPSec.Web.ViewModels
+﻿@using SAPSec.Web.ViewModels.Measures
 
 @model MeasureInfoViewModel
