+++ SAPSec.Web/Views/Shared/_MeasureYearByYearChart.cshtml
@@ -1,6 +1,7 @@
 ﻿@using SAPSec.Web.Constants
 @using System.Text.Json
-@using SAPSec.Web.ViewModels
+@using SAPSec.Web.ViewModels.Measures
+@using SAPSec.Core.Features.Measures
 @using SAPSec.Core.Features.Ks4HeadlineMeasures.UseCases
 
 @model YearByYearSubMeasureViewModel
@@ -35,6 +36,6 @@
             data-show-legend="true"
             data-show-datalabels="false"
             data-chart='@Html.Raw(JsonSerializer.Serialize(chartData))'
-            @if (Model.Colors is not null) { <text>data-colors='@Html.Raw(JsonSerializer.Serialize(Model.Colors))'</text> }>
+            @if (Model.MeasureInfo.YearByYearColors is not null) { <text>data-colors='@Html.Raw(JsonSerializer.Serialize(Model.MeasureInfo.YearByYearColors))'</text> }>
     </canvas>
 </div>
